resolve cyclic dependency by moving GraphHelper to gremlin-test
