+++ tinkergraph-gremlin/src/test/java/org/apache/tinkerpop/gremlin/tinkergraph/structure/TinkerGraphTest.java
@@ -20,6 +20,7 @@ package org.apache.tinkerpop.gremlin.tinkergraph.structure;
 
 import org.apache.commons.configuration.BaseConfiguration;
 import org.apache.commons.configuration.Configuration;
+import org.apache.tinkerpop.gremlin.GraphHelper;
 import org.apache.tinkerpop.gremlin.TestHelper;
 import org.apache.tinkerpop.gremlin.process.traversal.P;
 import org.apache.tinkerpop.gremlin.process.traversal.dsl.graph.GraphTraversalSource;
@@ -68,6 +69,7 @@ import java.util.function.Consumer;
 import java.util.function.Supplier;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotSame;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.mock;
 
@@ -591,6 +593,36 @@ public class TinkerGraphTest {
                 assertEquals(102, IteratorUtils.count(readX.edges()));
             }
         }
+    }
+
+    @Test
+    public void shouldCloneTinkergraph() {
+        final TinkerGraph original = TinkerGraph.open();
+        final TinkerGraph clone = TinkerGraph.open();
+
+        final Vertex marko = original.addVertex("name", "marko", "age", 29);
+        final Vertex stephen = original.addVertex("name", "stephen", "age", 35);
+        marko.addEdge("knows", stephen);
+        GraphHelper.cloneElements(original, clone);
+
+        final Vertex michael = clone.addVertex("name", "michael");
+        michael.addEdge("likes", marko);
+        michael.addEdge("likes", stephen);
+        clone.traversal().V().property("newProperty", "someValue").toList();
+        clone.traversal().E().property("newProperty", "someValue").toList();
+
+        assertEquals("original graph should be unchanged", new Long(2), original.traversal().V().count().next());
+        assertEquals("original graph should be unchanged", new Long(1), original.traversal().E().count().next());
+        assertEquals("original graph should be unchanged", new Long(0), original.traversal().V().has("newProperty").count().next());
+
+        assertEquals("cloned graph should contain new elements", new Long(3), clone.traversal().V().count().next());
+        assertEquals("cloned graph should contain new elements", new Long(3), clone.traversal().E().count().next());
+        assertEquals("cloned graph should contain new property", new Long(3), clone.traversal().V().has("newProperty").count().next());
+        assertEquals("cloned graph should contain new property", new Long(3), clone.traversal().E().has("newProperty").count().next());
+
+        assertNotSame("cloned elements should reference to different objects",
+            original.traversal().V().has("name", "stephen").next(),
+            clone.traversal().V().has("name", "stephen").next());
     }
 
     /**
